Refactoring chunks of premiere_get_referenced_media into seperate files for testing
