+++ src/asset_folder_importer.cfg
@@ -8,11 +8,10 @@ database_password={postgres-pass}
 start_path=/srv/Multimedia2/Media Production/Assets
 
 #Path where prelude projects live
-prelude_home=/srv/projectfiles/Projects
+prelude_home=/path/to/premiere_projects
 
 #Path where premiere projects live
-#premiere_home=/srv/projectfiles/Projects
-premiere_home=/Volumes/Project_Files_MultiMedia/Projects
+premiere_home=/path/to/premiere_projects
 
 #How to connect to the Vidispine database
 vs_host={vidispine-host}
@@ -31,3 +30,6 @@ vs_import_category=Rushes
 #housekeeping
 #purge system messages after this number of days
 system_message_purge_time = 5
+
+#if this is set, notify Sentry of errors
+sentry_dsn={my-dsn}
