+++ src/asset_folder_importer/premiere_get_referenced_media/Exceptions.py
@@ -1,0 +1,5 @@
+class InvalidDataError(StandardError):
+    pass
+
+class NoMediaError(StandardError):
+    pass
