+++ src/setup.py
@@ -10,6 +10,7 @@ setup(name='gnm-assetsweeper',
       packages=['asset_folder_importer',
                 'asset_folder_importer.asset_folder_sweeper',
                 'asset_folder_importer.asset_folder_vsingester',
+                'asset_folder_importer.premiere_get_referenced_media'
                 'asset_folder_importer.fix_unattached_media',
                 'asset_folder_importer.metadata_templates',
                 'asset_folder_importer.pluto',
